requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
asyncio>=3.4.3
aiohttp>=3.9.1
python-dateutil>=2.8.2
better-profanity>=0.7.0
praw>=7.7.1 